+++ test/index.js
@@ -1,11 +1,23 @@
 const mocha = require("mocha");
 const chai = require("chai");
 const expect = chai.expect;
+const inputData = require("../src/input-data");
+const stringValidation = require("../src/string-validation");
+const ipValues = inputData.getIpFour();
 
-describe("Description", function()
+describe("Validator Comparisons", function()
 {
-	it("Test Case", function()
+	describe("IP Address V4", function()
 	{
-		expect(true).to.be.true;
+		it("Valid", function()
+		{
+			stringValidation.compareIpFour(ipValues.valid, true);
+		});
+		
+		it("Invalid", function()
+		{
+			stringValidation.compareIpFour(ipValues.invalid, false);
+		});
+		
 	});
 });